+++ lib/ktoast.dart
@@ -1,3 +1,0 @@
-library ktoast;
-
-export 'src/toast.dart' show showToast, KToast, ToastPosition;