+++ lib/oktoast.dart
@@ -1,0 +1,3 @@
+library oktoast;
+
+export 'src/toast.dart' show showToast, OKToast, ToastPosition;